Moved catch.hpp to the tests subdirectory
